pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ct-collection-threshold-learning"
version = "0.1.0"
description = "Clinical Trial Data Analysis Pipeline with MOA classification, in-silico simulation, and threshold learning"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "alembic>=1.12",
    "pandas>=2.0",
    "numpy>=1.24",
    # HTTP
    "httpx>=0.25",
    # Visualization
    "plotly>=5.18",
    "kaleido>=0.2",
    # ML/Stats
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "statsmodels>=0.14",
    # Genomics
    "gseapy>=1.0",
    "anndata>=0.10",
    # Backend API
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "httpx",  # for FastAPI TestClient
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = [
    "connectors*",
    "database*",
    "moa_classification*",
    "analysis*",
    "visualization*",
    "api*",
    "config*",
]
